Moved ContentSerializer class to ContentSerializers directory
