+++ Distroir.CustomSDKLauncher.Core/Managers/ContentSerializers/ContentSerializer.cs
@@ -15,12 +15,8 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
-namespace Distroir.CustomSDKLauncher.Core.Managers
+namespace Distroir.CustomSDKLauncher.Core.Managers.ContentSerializers
 {
     //TODO: Rename Load to Deserialize and Load to Serialize
     public abstract class ContentSerializer<T>
